backend/requirements.txt

fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
email-validator
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
cryptography==42.0.0
email-validator==2.1.0
firebase-admin==6.4.0
google-cloud-firestore==2.14.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
requests==2.31.0
